chore(api-builder): rename linkDevGuide to linkDocs
Rename the specialized API docs `@linkDevGuide` tag to `@linkDocs`
since it is meant to be used to refer to all angular.io docs, not just
the developer guide.
