+++ tools/api-builder/links-package/index.js
@@ -3,7 +3,7 @@ var Package = require('dgeni').Package;
 module.exports = new Package('links', [])
 
 .factory(require('./inline-tag-defs/link'))
-.factory(require('./inline-tag-defs/linkDevGuide'))
+.factory(require('./inline-tag-defs/linkDocs'))
 .factory(require('./inline-tag-defs/example'))
 .factory(require('./inline-tag-defs/exampleTabs'))
 .factory(require('dgeni-packages/links/services/getAliases'))
@@ -12,9 +12,9 @@ module.exports = new Package('links', [])
 .factory(require('./services/parseArgString'))
 .factory(require('./services/getApiFragmentFileName'))
 
-.config(function(inlineTagProcessor, linkInlineTagDef, linkDevGuideInlineTagDef, exampleInlineTagDef, exampleTabsInlineTagDef) {
+.config(function(inlineTagProcessor, linkInlineTagDef, linkDocsInlineTagDef, exampleInlineTagDef, exampleTabsInlineTagDef) {
   inlineTagProcessor.inlineTagDefinitions.push(linkInlineTagDef);
-  inlineTagProcessor.inlineTagDefinitions.push(linkDevGuideInlineTagDef);
+  inlineTagProcessor.inlineTagDefinitions.push(linkDocsInlineTagDef);
   inlineTagProcessor.inlineTagDefinitions.push(exampleInlineTagDef);
   inlineTagProcessor.inlineTagDefinitions.push(exampleTabsInlineTagDef);
 });
